FinalProject/requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
pyFirmata==1.1.0
pyserial==3.5
